move all code responsible for testing our own code into ./tests
